requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
google-generativeai==0.8.3
chromadb==0.5.20
pypdf==5.1.0
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
tiktoken==0.8.0
httpx==0.27.2
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
